Revert "test: try double"
This reverts commit 76f2889d4c94327e0b3625d01220971bf5a47844.

+++ src/__init__.py
@@ -84,7 +84,7 @@ def interpolate(
     if grid_data.ndim == 3:
         grid_data = cp.ascontiguousarray(grid_data[:, :, :, cp.newaxis])
     nf = grid_data.shape[3]
-    result = cp.zeros((pos.shape[0], nf), dtype=cp.float64)
+    result = cp.zeros((pos.shape[0], nf), dtype=cp.float32)
     interpolateField(pos, grid_data, result, L, gradient, gradient_direction, kernel)
     return result
